권한그룹 추가기능 구현
